+++ lib/mokio/concerns/controllers/common.rb
@@ -52,6 +52,8 @@ module Mokio
           load_and_authorize_resource
         end
 
+
+
         #
         # Index action renders <b>json</b> with parameters to jquery.datatables
         #