+++ lib/mokio/concerns/controllers/contacts.rb
@@ -18,6 +18,11 @@ module Mokio
           obj.build_contact_template
         end
 
+        def set_breadcrumbs_prefix
+          @breadcrumbs_prefix = "content_management"
+          @breadcrumbs_prefix_link = "contents"
+        end
+
         private
           #
           # Never trust parameters from the scary internet, only allow the white list through.