+++ lib/mokio/concerns/controllers/external_scripts.rb
@@ -10,6 +10,11 @@ module Mokio
         included do
         end
 
+        def set_breadcrumbs_prefix
+          @breadcrumbs_prefix = "settings"
+          @breadcrumbs_prefix_link = ""
+        end
+
         private
           #
           # Never trust parameters from the scary internet, only allow the white list through.