+++ lib/mokio/concerns/controllers/users.rb
@@ -13,20 +13,25 @@ module Mokio
         end
 
         #
-        # Overriten new from CommonController#new (Mokio::Concerns::Controllers::Common)
+        # Overwritten new from CommonController#new (Mokio::Concerns::Controllers::Common)
         #
         def new
           @user = Mokio::User.new
         end
 
         #
-        # Overriten new from CommonController#update (Mokio::Concerns::Controllers::Common)
+        # Overwritten new from CommonController#update (Mokio::Concerns::Controllers::Common)
         #
         def update
-          @current = current_user 
+
+          # raise current_user.inspect
+          @current = current_user
+
           respond_to do |format|
             if obj.update(obj_params)
-              sign_in(obj, :bypass => true)
+              if obj.id == @current.id      #only when password os changed for current user, sign in (with updated password) is required
+                sign_in(obj, :bypass => true)
+              end
               if !params[:save_and_new].blank?
                 format.html { redirect_to obj_new_url(@obj_class.new), notice: CommonTranslation.updated(obj) }
                 format.json { head :no_content }
@@ -49,21 +54,16 @@ module Mokio
           # Request Referer for redirect user in the place where it was before the change password
           session[:return_to] = request.referer
           @password_only = true
-          @user = current_user
-          obj = @user
         end
 
         #
         # Updating password for user
         #
         def update_password
-          # @current = current_user
-          @user = current_user
-          obj = @user
           @user.only_password = true
           respond_to do |format|
             if @user.update(user_params)
-              sign_in(@user, :bypass => true) #I18n.t("prices.quotation_not_created", title: l(@date))
+              sign_in(@user, :bypass => true)
 
               format.html { redirect_to session[:return_to] ,notice: I18n.t("users.password_updated") }
               # format.json { render action: 'edit_password', status: :updated}
@@ -73,6 +73,11 @@ module Mokio
               format.json { render json: @user.errors, status: :unprocessable_entity }
             end
           end
+        end
+
+        def set_breadcrumbs_prefix
+          @breadcrumbs_prefix = "settings"
+          @breadcrumbs_prefix_link = ""
         end
 
         private