+++ lib/mokio/concerns/models/menu.rb
@@ -233,6 +233,13 @@ module Mokio
 
           type.tableize
         end
+
+        #
+        # Specify what's showed in breadcrumb
+        #
+        def breadcrumb_name
+          name
+        end
       end
     end
   end