+++ lib/mokio/concerns/models/module_position.rb
@@ -37,6 +37,13 @@ module Mokio
           true
         end
 
+      #
+      # Specify what's showed in breadcrumb
+      #
+      def breadcrumb_name
+        name
+      end
+
       end
     end
   end